requirements.txt
h5py>=3.7.0
keras>=2.11.0
matplotlib>=3.6.2
numpy>=1.23.5
Pillow>=9.3.0
tensorflow>=2.11.0

